Remove old test
